Rename user-enters-information -> i-can-enter-information
