+++ cypress/integration/individual-pages/activity.spec.ts
@@ -1,3 +1,4 @@
+import { asAMaritimePleasureBeaconOwner } from "../i-can-enter-information";
 import {
   andIHaveEnteredNoInformation,
   andIHaveSelected,
@@ -8,7 +9,6 @@ import {
   whenIClickContinue,
   whenIClickOnTheErrorSummaryLinkContaining,
 } from "../selectors-and-assertions.spec";
-import { asAMaritimePleasureBeaconOwner } from "../user-enters-information.spec";
 
 describe("As a beacon owner, I want to submit the primary activity for my beacon", () => {
   const otherActivitySelector = "#other-activity";
