+++ cypress/integration/individual-pages/purpose.spec.ts
@@ -1,10 +1,10 @@
+import { asAnAviationBeaconOwner } from "../i-can-enter-information";
 import {
   andIHaveEnteredNoInformation,
   requiredFieldErrorMessage,
   thenIShouldSeeAnErrorMessageThatContains,
   whenIClickContinue,
 } from "../selectors-and-assertions.spec";
-import { asAnAviationBeaconOwner } from "../user-enters-information.spec";
 
 describe("As a beacon owner, I want to submit the purpose for my beacon", () => {
   it("displays an error if no beacon use purpose is selected", () => {
