+++ cypress/integration/single-beacon-owner/maritime.spec.ts
@@ -1,4 +1,5 @@
 import { PageURLs } from "../../../src/lib/urls";
+import { givenIHaveEnteredMyBeaconDetails } from "../i-can-enter-information";
 import {
   andIClickContinue,
   givenIHaveSelected,
@@ -10,7 +11,6 @@ import {
   thenTheUrlShouldContain,
   whenIClickBack,
 } from "../selectors-and-assertions.spec";
-import { givenIHaveEnteredMyBeaconDetails } from "../user-enters-information.spec";
 
 describe("As a maritime beacon owner,", () => {
   it("I can register my beacon for pleasure purposes", () => {
